requirements.txt
anyio==4.6.2.post1
asgiref==3.8.1
async-timeout==4.0.3
channels==4.1.0
channels-redis==4.2.0
click==8.1.7
Django==5.1.2
django-cors-headers==4.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
exceptiongroup==1.2.2
h11==0.14.0
httptools==0.6.4
idna==3.10
msgpack==1.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.1.1
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1